+++ neuron/src/app/Neuron/CLI/Types.hs
@@ -13,6 +13,7 @@ module Neuron.CLI.Types
     SearchBy (..),
     SearchCommand (..),
     OpenCommand (..),
+    QueryCommand (..),
     RibConfig (..),
     commandParser,
   )
@@ -67,6 +68,13 @@ data OpenCommand = OpenCommand
   }
   deriving (Eq, Show)
 
+data QueryCommand = QueryCommand
+  { -- Use cache instead of building the zettelkasten from scratch
+    cached :: Bool,
+    query :: Either (Some Q.ZettelQuery) (Some Q.GraphQuery)
+  }
+  deriving (Eq, Show)
+
 data Command
   = -- | Create a new zettel file
     New NewCommand
@@ -75,7 +83,7 @@ data Command
   | -- | Search a zettel by title
     Search SearchCommand
   | -- | Run a query against the Zettelkasten
-    Query (Either (Some Q.ZettelQuery) (Some Q.GraphQuery))
+    Query QueryCommand
   | -- | Delegate to Rib's command parser
     Rib RibConfig
 
@@ -149,38 +157,46 @@ commandParser defaultNotesDir today = do
           <|> fmap
             (OpenCommand . Some . R.Route_Zettel)
             (option zettelIDReader (long "id" <> help "Open the zettel HTML page" <> metavar "ID"))
-    queryCommand =
-      fmap Query $
-        ( fmap
-            Left
-            ( fmap (Some . flip Q.ZettelQuery_ZettelByID Nothing) (option zettelIDReader (long "id"))
-                <|> fmap (\x -> Some $ Q.ZettelQuery_ZettelsByTag x Nothing def) (many (mkTagPattern <$> option str (long "tag" <> short 't')))
-                <|> option queryReader (long "uri" <> short 'u')
+    queryCommand = do
+      cached <- switch (long "cached" <> help "Use cached zettelkasten graph (faster)")
+      query <-
+        fmap
+          Left
+          ( fmap
+              (Some . flip Q.ZettelQuery_ZettelByID Nothing)
+              (option zettelIDReader (long "id"))
+              <|> fmap
+                (\x -> Some $ Q.ZettelQuery_ZettelsByTag x Nothing def)
+                (many (mkTagPattern <$> option str (long "tag" <> short 't')))
+              <|> option queryReader (long "uri" <> short 'u')
+          )
+          <|> fmap
+            Right
+            ( fmap
+                (const $ Some $ Q.GraphQuery_Id)
+                ( switch $
+                    long "graph" <> help "Get the entire zettelkasten graph as JSON"
+                )
+                <|> fmap
+                  (Some . Q.GraphQuery_BacklinksOf Nothing)
+                  ( option
+                      zettelIDReader
+                      ( long "backlinks-of"
+                          <> help "Get backlinks to the given zettel ID"
+                          <> metavar "ID"
+                      )
+                  )
+                <|> fmap
+                  (Some . Q.GraphQuery_BacklinksOf (Just C.Folgezettel))
+                  ( option
+                      zettelIDReader
+                      ( long "uplinks-of"
+                          <> help "Get uplinks to the given zettel ID"
+                          <> metavar "ID"
+                      )
+                  )
             )
-            <|> fmap
-              Right
-              (fmap (const $ Some $ Q.GraphQuery_Id) $ switch (long "graph" <> help "Get the entire zettelkasten graph as JSON"))
-            <|> fmap
-              Right
-              ( fmap (Some . Q.GraphQuery_BacklinksOf Nothing) $
-                  option
-                    zettelIDReader
-                    ( long "backlinks-of"
-                        <> help "Get backlinks to the given zettel ID"
-                        <> metavar "ID"
-                    )
-              )
-            <|> fmap
-              Right
-              ( fmap (Some . Q.GraphQuery_BacklinksOf (Just C.Folgezettel)) $
-                  option
-                    zettelIDReader
-                    ( long "uplinks-of"
-                        <> help "Get uplinks to the given zettel ID"
-                        <> metavar "ID"
-                    )
-              )
-        )
+      pure $ Query $ QueryCommand {..}
     searchCommand = do
       searchBy <-
         bool SearchByTitle SearchByContent