+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - Styling fixes
     - Fix incorrect body font (esp. on Windows)
     - Fix task list checkbox styling (#233)
+- CLI
+  - Add `--id` and `--search` options to `open` command to open given zettel ID or search page respectively  (#317)
 
 ## 0.6.0.0