+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Zettel format
   - Add `unlisted` metadata property to hide a zettel from z-index (#318)
 - CLI
+  - Faster querying: add `--cached` option to `neuron query`, to run faster using the cache. To keep the cache up to date, make sure that `neuron rib` is running.
   - Add `--id` and `--search` options to `open` command to open given zettel ID or search page respectively  (#317)
 
 ## 0.6.0.0