+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Fix task list checkbox styling (#233)
 - Zettel format
   - Add `unlisted` metadata property to hide a zettel from z-index (#318)
+  - Markdown: support for [fancy lists](https://github.com/jgm/commonmark-hs/blob/master/commonmark-extensions/test/fancy_lists.md) (#335)
 - CLI
   - Faster querying: add `--cached` option to `neuron query`, to run faster using the cache. To keep the cache up to date, make sure that `neuron rib` is running.
   - Add `--id` and `--search` options to `open` command to open given zettel ID or search page respectively  (#317)